Saving last changes

+++ src/bigNumbers.java
@@ -1,4 +1,4 @@
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.Scanner;
 
 public class bigNumbers {
@@ -9,17 +9,16 @@ public class bigNumbers {
 
 
         System.out.println("Enter a value for a");
-        BigInteger a = scanner.nextBigInteger();
+        BigDecimal a = scanner.nextBigDecimal();
         System.out.println("Enter a value for b");
-        BigInteger b = scanner.nextBigInteger();
-
+        BigDecimal b = scanner.nextBigDecimal();
 
 
             // when a is bigger than b, compareTo method returns 1
         if (a.compareTo(b) > 0) {
             System.out.println("A is greater than B");
             // when a is less than b, compareTo method returns -1
-        } else if (a.compareTo(b) < 0){
+        } else if (a.compareTo(b) < 0) {
             System.out.println("A is less than B");
             // when a is equal to b, compare to method returns 0
         } else if (a.compareTo(b) == 0) {
